remove download csv from header
