+++ apps/nowcasting-app/components/layout/header/index.tsx
@@ -4,12 +4,9 @@ import { OCFlogo } from "../../icons/logo";
 import Link from "next/link";
 import { Menu } from "@headlessui/react";
 import { getViewTitle, VIEWS } from "../../../constant";
-import { Dispatch, SetStateAction, useState } from "react";
+import { Dispatch, SetStateAction } from "react";
 import { ExternalLinkIcon } from "../../icons/icons";
-import useGlobalState from "../../helpers/globalState";
 import { CombinedData } from "../../types";
-import { downloadNationalCsv } from "../../helpers/csvDownload";
-import { CSVDownloadModal, CSVColumn } from "./csvDownloadModal";
 
 type HeaderLinkProps = {
   url: string;
@@ -93,93 +90,70 @@ const Header: React.FC<HeaderProps> = ({
   isLoggedIn = true,
   combinedData = null
 }) => {
-  const [showDownloadModal, setShowDownloadModal] = useState(false);
-  const [nHourForecast] = useGlobalState("nHourForecast");
-  const canDownloadCsv = Boolean(isLoggedIn && combinedData && view !== VIEWS.SOLAR_SITES);
-
-  const handleDownload = (selectedColumns: CSVColumn[]) => {
-    downloadNationalCsv(combinedData, selectedColumns, nHourForecast);
-  };
-
   return (
-    <>
-      <header className="h-16 text-white text-right sm:px-4 bg-black flex absolute top-0 w-full overflow-y-visible p-1 text-sm items-center z-30">
-        <div className="flex-grow-0 -mt-0.5 flex-shrink-0">
-          <a
-            className="flex h-8 self-center w-auto"
-            target="_blank"
-            href="https://quartz.solar/"
-            rel="noreferrer"
-          >
-            <img src="/QUARTZSOLAR_LOGO_ICON.svg" alt="quartz_logo" className="h-8 w-auto" />
-          </a>
+    <header className="h-16 text-white text-right sm:px-4 bg-black flex absolute top-0 w-full overflow-y-visible p-1 text-sm items-center z-30">
+      <div className="flex-grow-0 -mt-0.5 flex-shrink-0">
+        <a
+          className="flex h-8 self-center w-auto"
+          target="_blank"
+          href="https://quartz.solar/"
+          rel="noreferrer"
+        >
+          <img src="/QUARTZSOLAR_LOGO_ICON.svg" alt="quartz_logo" className="h-8 w-auto" />
+        </a>
+      </div>
+      <div className="p-1 mt-0.5 mb-1.5 items-end flex flex-col">
+        <a
+          className="flex h-6 w-auto"
+          target="_blank"
+          href="https://quartz.solar/"
+          rel="noreferrer"
+        >
+          <img
+            src="/QUARTZSOLAR_LOGO_TEXTONLY_WHITE.svg"
+            alt="quartz_logo"
+            className="h-8 w-auto"
+          />
+        </a>
+        <div className="mr-[6px] flex items-center">
+          <span className="block mr-[1px] font-light tracking-wide text-[10px]">powered by</span>
+          <OCFlogo />
         </div>
-        <div className="p-1 mt-0.5 mb-1.5 items-end flex flex-col">
-          <a
-            className="flex h-6 w-auto"
-            target="_blank"
-            href="https://quartz.solar/"
-            rel="noreferrer"
-          >
-            <img
-              src="/QUARTZSOLAR_LOGO_TEXTONLY_WHITE.svg"
-              alt="quartz_logo"
-              className="h-8 w-auto"
+      </div>
+      <div className="grow text-center inline-flex px-2 sm:px-8 gap-2 sm:gap-5 items-center">
+        {isLoggedIn && (
+          <Menu>
+            <HeaderLink
+              url="/"
+              view={VIEWS.FORECAST}
+              currentView={view}
+              setViewFunc={setView}
+              text={getViewTitle(VIEWS.FORECAST)}
             />
-          </a>
-          <div className="mr-[6px] flex items-center">
-            <span className="block mr-[1px] font-light tracking-wide text-[10px]">powered by</span>
-            <OCFlogo />
-          </div>
+            <HeaderLink
+              url="/"
+              view={VIEWS.SOLAR_SITES}
+              currentView={view}
+              setViewFunc={setView}
+              text={getViewTitle(VIEWS.SOLAR_SITES)}
+              disabled={isProduction}
+            />
+            <HeaderLink
+              url="/"
+              view={VIEWS.DELTA}
+              currentView={view}
+              setViewFunc={setView}
+              text={getViewTitle(VIEWS.DELTA)}
+            />
+          </Menu>
+        )}
+      </div>
+      <div className="flex items-center gap-2">
+        <div className="py-1">
+          {isLoggedIn && <ProfileDropDown view={view} combinedData={combinedData} />}
         </div>
-        <div className="grow text-center inline-flex px-2 sm:px-8 gap-2 sm:gap-5 items-center">
-          {isLoggedIn && (
-            <Menu>
-              <HeaderLink
-                url="/"
-                view={VIEWS.FORECAST}
-                currentView={view}
-                setViewFunc={setView}
-                text={getViewTitle(VIEWS.FORECAST)}
-              />
-              <HeaderLink
-                url="/"
-                view={VIEWS.SOLAR_SITES}
-                currentView={view}
-                setViewFunc={setView}
-                text={getViewTitle(VIEWS.SOLAR_SITES)}
-                disabled={isProduction}
-              />
-              <HeaderLink
-                url="/"
-                view={VIEWS.DELTA}
-                currentView={view}
-                setViewFunc={setView}
-                text={getViewTitle(VIEWS.DELTA)}
-              />
-            </Menu>
-          )}
-        </div>
-        <div className="flex items-center gap-2">
-          <div className="py-1">
-            {isLoggedIn && (
-              <ProfileDropDown
-                canDownloadCsv={canDownloadCsv}
-                onDownloadCsv={() => setShowDownloadModal(true)}
-              />
-            )}
-          </div>
-        </div>
-      </header>
-
-      <CSVDownloadModal
-        isOpen={showDownloadModal}
-        onClose={() => setShowDownloadModal(false)}
-        onDownload={handleDownload}
-        nHourForecast={nHourForecast}
-        view={view}
-      />
-    </>
+      </div>
+    </header>
   );
 };
